Add CCP envvar values to rendering context
I am creating this CR for discussion.

I have the following in my container definition:

    env:
      - name: CCP_NODE_NAME
        valueFrom:
          fieldRef:
            fieldPath: spec.nodeName

This is to be able to know the node name from my pod
container.

I also need to set some variable to the node name in my
service's configuration file. So I'd like to use this in
this configuration file:

hostname = {{ CCP_NODE_NAME }}

But for this to work the Jinja rendering context should
include the CCP_NODE_NAME, set to the correct value.

Hence this patch.

Change-Id: I293bb3aa113883c02f2e738f9d74291bf2f23d95